// src/app/models/error-messages.ts
export enum ErrorMessages {
  INVALID_EMAIL = 'This email address is invalid',
  REQUIRED = 'This field is required',
  INVALID_PASSWORD = 'Password must have at least 8 characters',
  INVALID_VERIFY_PASSWORD = 'Passwords don\'t match',
  INVALID_PHONE = 'Enter a 10 digit phone number',
  INVALID_PIN = 'Enter a 6 digit pin code'
}